Small improvements.
Coypright header, docs, formatting & moved it into the core project.
